refactor message resanding
