+++ consensus/tendermint/backend/backend.go
@@ -19,7 +19,6 @@ package backend
 import (
 	"context"
 	"crypto/ecdsa"
-	"errors"
 	"fmt"
 	"math/big"
 	"sync"
@@ -212,212 +211,6 @@ func peersToString(ps []common.Address) string {
 	}
 
 	return peersStr
-}
-
-func getPeerKeys(psMap map[common.Address]consensus.Peer) []common.Address {
-	ps := make([]common.Address, 0, len(psMap))
-	for k := range psMap {
-		ps = append(ps, k)
-	}
-	return ps
-}
-
-type message struct {
-	hash    common.Hash
-	payload []byte
-}
-
-type peerError struct {
-	error
-	addr common.Address
-}
-
-func (sb *Backend) sendToPeer(ctx context.Context, addr common.Address, hash common.Hash, payload []byte, p consensus.Peer) chan error {
-	ms, ok := sb.recentMessages.Get(addr)
-	errCh := make(chan error, 1)
-	var m *lru.ARCCache
-	if ok {
-		m, _ = ms.(*lru.ARCCache)
-		if _, k := m.Get(hash); k {
-			// This peer had this event, skip it
-			sb.logger.Trace("inner sender loop. message sent earlier", "peer", addr.Hex(), "msg", hash.Hex())
-			errCh <- nil
-			return errCh
-		}
-	} else {
-		m, _ = lru.NewARC(inmemoryMessages)
-	}
-
-	go func(ctx context.Context, p consensus.Peer, m *lru.ARCCache) {
-		ticker := time.NewTicker(retryInterval * time.Millisecond)
-		defer ticker.Stop()
-
-		var err error
-		var try int
-
-	SenderLoop:
-		for {
-			select {
-			case <-ticker.C:
-				try++
-
-				if err = p.Send(tendermintMsg, payload); err != nil {
-					err = peerError{errors.New("error while sending tendermintMsg message to the peer: " + err.Error()), addr}
-
-					sb.logger.Trace("inner sender loop. error", "try", try, "peer", addr.Hex(), "msg", hash.Hex(), "err", err.Error())
-				} else {
-					err = nil
-
-					sb.logger.Trace("inner sender loop. success", "try", try, "peer", addr.Hex(), "msg", hash.Hex())
-					break SenderLoop
-				}
-			case <-ctx.Done():
-				err = peerError{errors.New("error while sending tendermintMsg message to the peer: " + ctx.Err().Error()), addr}
-				break SenderLoop
-			}
-		}
-
-		if err == nil {
-			m.Add(hash, true)
-			sb.recentMessages.Add(addr, m)
-		}
-
-		errCh <- err
-	}(ctx, p, m)
-
-	return errCh
-}
-
-func (sb *Backend) ReSend(ctx context.Context, numberOfWorkers int) {
-	workers := sync.WaitGroup{}
-	workers.Add(numberOfWorkers)
-
-	for i := 0; i < numberOfWorkers; i++ {
-		go func(ctx context.Context) {
-			sb.workerSendLoop(ctx)
-			workers.Done()
-		}(ctx)
-	}
-
-	workers.Wait()
-}
-
-func (sb *Backend) workerSendLoop(ctx context.Context) {
-	for {
-		select {
-		case msgToPeers := <-sb.resend:
-			sb.trySend(ctx, msgToPeers)
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-func (sb *Backend) sendToResendCh(ctx context.Context, m messageToPeers) {
-	select {
-	case <-ctx.Done():
-		return
-	case sb.resend <- m:
-		//sent to channel
-	}
-}
-
-func (sb *Backend) trySend(ctx context.Context, msgToPeers messageToPeers) {
-	if int(time.Since(msgToPeers.startTime).Seconds()) > TTL {
-		sb.logger.Trace("worker loop. TTL expired", "msg", msgToPeers)
-		return
-	}
-
-	if time.Since(msgToPeers.lastTry).Truncate(time.Millisecond).Nanoseconds()/int64(time.Millisecond) < retryInterval &&
-		time.Since(msgToPeers.startTime).Truncate(time.Millisecond).Nanoseconds()/int64(time.Millisecond) > retryInterval {
-
-		//too early for resend
-		sb.sendToResendCh(ctx, msgToPeers)
-
-		return
-	}
-
-	m := make(map[common.Address]struct{})
-	for _, p := range msgToPeers.peers {
-		m[p] = struct{}{}
-	}
-
-	ps, notConnected := sb.broadcaster.FindPeers(m)
-	notConnectedPeersBeforeSending := len(notConnected)
-	if len(notConnected) > 0 {
-		peersStr := fmt.Sprintf("peers %d: ", len(notConnected))
-		for _, p := range notConnected {
-			peersStr = fmt.Sprintf("%s%s ", peersStr, p.Hex())
-		}
-
-		sb.logger.Trace("worker loop. peers still not connected", "peers", peersStr, "msgHash", msgToPeers.msg.hash.String())
-	}
-
-	if sb.broadcaster != nil && len(ps) > 0 {
-		sb.logger.Trace("worker loop. resend to connected peers", "msg", msgToPeers.msg.hash.String(), "peers", peersToString(getPeerKeys(ps)))
-		var errChs []chan error
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, TTL*time.Second)
-		defer cancel()
-
-		for addr, p := range ps {
-			errCh := sb.sendToPeer(ctx, addr, msgToPeers.msg.hash, msgToPeers.msg.payload, p)
-			errChs = append(errChs, errCh)
-		}
-
-		wg := sync.WaitGroup{}
-		wg.Add(len(ps))
-		notConnectedCh := make(chan common.Address, len(ps))
-
-		for _, errCh := range errChs {
-			go func(errCh chan error) {
-				err := <-errCh
-				if err != nil {
-					pe, ok := err.(peerError)
-					if ok {
-						notConnectedCh <- pe.addr
-
-						sb.logger.Error(pe.Error(), "peer", pe.addr)
-					}
-				}
-
-				close(errCh)
-				wg.Done()
-			}(errCh)
-		}
-
-		wg.Wait()
-		close(notConnectedCh)
-
-		for addr := range notConnectedCh {
-			notConnected = append(notConnected, addr)
-		}
-	}
-
-	notConnectedPeersAfterSendingWithErrors := len(notConnected)
-
-	if int(time.Since(msgToPeers.startTime).Seconds()) > TTL {
-		sb.logger.Trace("worker loop. TTL expired", "msg", msgToPeers)
-		return
-	}
-
-	if len(notConnected) > 0 {
-		msg := messageToPeers{
-			message{
-				msgToPeers.msg.hash,
-				msgToPeers.msg.payload,
-			},
-			notConnected,
-			msgToPeers.startTime,
-			time.Now(),
-		}
-
-		if notConnectedPeersBeforeSending != notConnectedPeersAfterSendingWithErrors {
-			sb.logger.Trace("worker loop. peers not connected and errors while sending", "msg", msg.String())
-		}
-
-		sb.sendToResendCh(ctx, msg)
-	}
 }
 
 // Commit implements tendermint.Backend.Commit
