+++ consensus/tendermint/backend/engine.go
@@ -491,7 +491,7 @@ func (sb *Backend) Start(chain consensus.ChainReader, currentBlock func() *types
 	sb.coreStarted = true
 
 	sb.resend = make(chan messageToPeers, 1024)
-	go sb.ReSend(ctx, 10)
+	sb.ReSend(ctx, 10)
 
 	return nil
 }
